Replace XULONG with XDWORD
The XULONG type (unsigned long) is 4 bytes on Windows but 8 bytes on
64-bit Linux/macOS, causing memory layout issues and serialization errors.
Replacing it with XDWORD (always 4 bytes) ensures consistent struct sizes
across all platforms.

Additional changes:
- Use size_t for memory and buffer sizes
- Fix pointer arithmetic signedness in array classes
- Update XObjectArray::RemoveObject return logic
- Remove legacy XULONG typedef definitions

+++ include/XSArray.h
@@ -293,7 +293,7 @@ public:
     /// @brief Moves an element from one position to another.
     void Move(T *i, T *n) {
         if (i >= m_Begin && i <= m_End && n >= m_Begin && n < m_End) {
-            int insertpos = i - m_Begin;
+            int insertpos = static_cast<int>(i - m_Begin);
             if (n < i) --insertpos;
             T tn = *n;
 #if VX_HAS_CXX11
@@ -439,13 +439,13 @@ protected:
 
     void XCopy(T *dest, T *start, T *end) {
         if (start == end) return;
-        int size = ((XBYTE *) end - (XBYTE *) start);
+        size_t size = static_cast<size_t>((XBYTE *) end - (XBYTE *) start);
         if (size > 0) memcpy(dest, start, size);
     }
 
     void XMove(T *dest, T *start, T *end) {
         if (start == end) return;
-        int size = ((XBYTE *) end - (XBYTE *) start);
+        size_t size = static_cast<size_t>((XBYTE *) end - (XBYTE *) start);
         if (size > 0) memmove(dest, start, size);
     }
