+++ tests/VxMath/SIMDFeatureTest.cpp
@@ -119,7 +119,7 @@ protected:
     }
 
     void* AllocAndTrack(size_t size, size_t alignment) {
-        void* ptr = VxNewAligned(size, alignment);
+        void* ptr = VxNewAligned(size, static_cast<int>(alignment));
         if (ptr) m_allocated.push_back(ptr);
         return ptr;
     }
